requirements.txt
faust-streaming==0.8.4
psycopg2-binary
